backend/requirements.txt
python-dotenv
sentence-transformers
fastembed
pdfplumber
langchain-text-splitters
qdrant-client
hdbscan
ollama
numpy
pandas
fastapi
uvicorn[standard]
